fix:字段类型错误

+++ src/main/java/com/dezhishen/domain/Artist.java
@@ -15,6 +15,6 @@ public class Artist {
     private String name;
     private String picUrl;
     private String img1v1Url;
-    private int img1v1;
+    private long img1v1;
     private String trans;
 }
